refactor: pr시 요구사항 수정

+++ src/main/java/com/skyhorsemanpower/chatService/review/data/vo/SearchAuctionReviewResponseVo.java
@@ -8,11 +8,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class SearchAuctionReviewResponseVo {
     private String reviewWriterUuid;
-    private Byte reviewRate;
+    private int reviewRate;
     private String reviewContent;
-
+    // 리뷰 작성자 uuid 담는 것은 타 api 활용하는거 확인 후 수정 예정
     @Builder
-    public SearchAuctionReviewResponseVo(String reviewWriterUuid, Byte reviewRate,
+    public SearchAuctionReviewResponseVo(String reviewWriterUuid, int reviewRate,
         String reviewContent) {
         this.reviewWriterUuid = reviewWriterUuid;
         this.reviewRate = reviewRate;
